requirements/utils-cpu.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
cmake==3.20.3
contourpy==1.1.1
cycler==0.12.1
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
hf-xet==1.1.7
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.1
numba==0.58.1
numpy==1.23.5
omegaconf==2.3.0
onnxruntime==1.20.1
openunmix==1.2.1
packaging==25.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
retrying==1.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.1
six==1.17.0
soundfile==0.13.1
soxr==0.3.7
submitit==1.5.3
sympy==1.13.3
threadpoolctl==3.5.0
tqdm==4.67.1
treetable==0.2.5
typing_extensions==4.13.2
urllib3==2.2.3
uv==0.8.7
wheel==0.45.1
transformers==4.56.0
zipp==3.20.2
